[Bug304438] Introduced TmfLocation

+++ org.eclipse.linuxtools.tmf/src/org/eclipse/linuxtools/tmf/request/TmfDataRequest.java
@@ -141,7 +141,7 @@ public class TmfDataRequest<T extends TmfData> {
 	/**
 	 * @return the index
 	 */
-	public int getIndex() {
+	public long getIndex() {
 		return fIndex;
 	}
